cooin-backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# HTTP Client for external API calls
httpx==0.25.2

# Validation & Serialization
email-validator==2.1.0
pydantic[email]==2.5.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0

# CORS is handled by FastAPI's built-in middleware
# No separate package needed

# Logging & Monitoring
structlog==23.2.0

# Background Tasks (optional for future use)
celery==5.3.4
redis==5.0.1

# File handling (for profile images, documents)
python-magic==0.4.27
pillow==10.1.0

# Date & Time utilities
python-dateutil==2.8.2

# Production WSGI server (alternative to uvicorn)
gunicorn==21.2.0